begin outlining action module
