+++ django_unicorn/actions/refresh.py
@@ -1,0 +1,10 @@
+from django_unicorn.components import Component
+from django_unicorn.views.actions.base import Action, ActionResult
+
+class Refresh(Action):
+    
+    action_type = "callMethod"
+    method_name = "$refresh"
+    
+    def apply(self, component: Component) -> ActionResult:
+        raise NotImplementedError()
