+++ django_unicorn/actions/reset.py
@@ -1,0 +1,10 @@
+from django_unicorn.components import Component
+from django_unicorn.views.actions.base import Action, ActionResult
+
+class Reset(Action):
+    
+    action_type = "callMethod"
+    method_name = "$reset"
+    
+    def apply(self, component: Component) -> ActionResult:
+        raise NotImplementedError()
