+++ django_unicorn/actions/sync_input.py
@@ -1,0 +1,9 @@
+from django_unicorn.components import Component
+from django_unicorn.views.actions.base import Action, ActionResult
+
+class SyncInput(Action):
+    
+    action_type = "syncInput"
+    
+    def apply(self, component: Component) -> ActionResult:
+        raise NotImplementedError()
