+++ django_unicorn/actions/validate.py
@@ -1,0 +1,10 @@
+from django_unicorn.components import Component
+from django_unicorn.views.actions.base import Action, ActionResult
+
+class Validate(Action):
+    
+    action_type = "callMethod"
+    method_name = "$validate"
+    
+    def apply(self, component: Component) -> ActionResult:
+        raise NotImplementedError()
