+++ django_unicorn/components/__init__.py
@@ -1,5 +1,5 @@
 from django_unicorn.components.mixins import ModelValueMixin
-from django_unicorn.components.unicorn_view import UnicornField, UnicornView
+from django_unicorn.components.unicorn_view import UnicornField, UnicornView, Component
 from django_unicorn.components.updaters import HashUpdate, LocationUpdate, PollUpdate
 from django_unicorn.typing import QuerySetType
