+++ django_unicorn/views/utils.py
@@ -7,9 +7,23 @@ from django.db.models import Model
 from django_unicorn.components import UnicornField, UnicornView
 from django_unicorn.decorators import timed
 from django_unicorn.typer import cast_value, create_queryset, get_type_hints, is_queryset
+from django_unicorn.utils import is_int
 
 logger = logging.getLogger(__name__)
 
+def sort_dict(d):
+    items = [
+        [k, v]
+        for k, v in sorted(
+            d.items(), key=lambda x: x[0] if not is_int(x[0]) else int(x[0])
+        )
+    ]
+
+    for item in items:
+        if isinstance(item[1], dict):
+            item[1] = sort_dict(item[1])
+
+    return dict(items)
 
 @timed
 def set_property_from_data(
